Simplify get_all_attrs implementation
Refactor the `get_all_attrs` function to use a single loop instead of two.
This change removes the need for a temporary array to store attribute
pointers, simplifying the code.

The attribute is now fetched and copied within the same iteration, making the
logic more direct and easier to follow.

This eliminates a Scan Build issue.

Signed-off-by: Simo Sorce <simo@redhat.com>