Improve object pool membership tracking
Use the `poolid` field to reliably track whether an object is part of the
object pool. A value of -1 now indicates an object is not in the pool.

This replaces the previous, less reliable method of checking the `slotid`.

This also ensures that duplicated key objects are correctly added to the pool
and adds a safeguard to prevent overwriting an existing pooled object during a
duplication operation.

Signed-off-by: Simo Sorce <simo@redhat.com>