Use sizeof for key template attribute count
The `p11prov_create_secret_key` function used a hardcoded value for the number
of attributes in the key template.

This is replaced with a compile-time calculation using `sizeof` to determine
the number of elements in the array. This makes the code more robust and
maintainable, as changes to the template array will not require manually
updating the count.

Signed-off-by: Simo Sorce <simo@redhat.com>