Fix NULL dereference in RSA signature verification
The check for a null signature context and a null signature was
previously combined in a single if statement.

If the context `sigctx` was NULL, the code in the condition would
lead to a null pointer dereference.
This splits the validation into two separate checks to ensure the
context exists before its members are accessed.

Fixes CID 638281

Signed-off-by: Simo Sorce <simo@redhat.com>