Fix memory leak in tpkey test
In the tpkey test, the special case for P-256 keys with the softhsm driver
would skip the test execution via a 'continue' statement.

The EVP_PKEY was allocated before this check but not freed, leading to a
memory leak in each skipped iteration. This change adds the required
EVP_PKEY_free() call.

Signed-off-by: Simo Sorce <simo@redhat.com>